requirements.txt
customtkinter==5.2.2
opencv_python==4.9.0.80
Pillow==10.3.0
pyttsx3==2.90
ultralytics==8.1.9
